check cached dependencies

+++ build_scripts/01-openssl.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 
-echo "Downloading openssl..."
-curl -fsS https://drop.bakatrouble.pw/_/openssl.tar.gz -o openssl.tar.gz
-tar -xf openssl.tar.gz
+cd ..
+
+if [ -f openssl/lib/Win32/static/Release/libssl.lib ]; then
+    echo "OpenSSL is already downloaded"
+else
+    echo "Downloading OpenSSL..."
+    curl -fsS https://drop.bakatrouble.pw/_/openssl.tar.gz -o openssl.tar.gz
+    tar -xf openssl.tar.gz
+fi
